requirements.txt
fastapi>=0.110
uvicorn>=0.27
jinja2>=3.1
python-multipart>=0.0.9
aiofiles>=23.2
docling
python-dotenv>=1.0
sqlalchemy>=2.0
requests>=2.31
stripe>=8.0
